+++ web/tests/navbar_alerts.test.cjs
@@ -152,16 +152,3 @@ test("should_show_server_upgrade_banner", ({override}) => {
     override(Date, "now", () => addDays(start_time, 8).getTime()); // Thursday 1/9/2024 10:00:00 AM (UTC+0)
     assert.equal(navbar_alerts.should_show_server_upgrade_banner(ls), true);
 });
-
-test("get_demo_organization_deadline_days_remaining", ({override}) => {
-    const start_time = new Date("2024-01-01T10:00:00.000Z"); // Wednesday 1/1/2024 10:00:00 AM (UTC+0)
-    override(Date, "now", () => start_time);
-
-    const demo_organization_scheduled_deletion_date = addDays(start_time, 7); // Wednesday 1/8/2024 10:00:00 AM (UTC+0)
-    override(
-        realm,
-        "demo_organization_scheduled_deletion_date",
-        Math.trunc(demo_organization_scheduled_deletion_date / 1000),
-    );
-    assert.equal(navbar_alerts.get_demo_organization_deadline_days_remaining(), 7);
-});
